refactor(health): move health check logic to a new module and simplify health endpoint
